requirements.txt
# VisionCore Enterprise Edition - Python Dependencies
# ========================================================
# 核心依赖 - 必须安装
# ========================================================

# AI推理框架
ultralytics>=8.0.0              # YOLOv8 模型推理（PC端）
torch>=2.0.0                    # PyTorch深度学习框架（PC端）
torchvision>=0.15.0             # PyTorch视觉库

# 图像处理
opencv-python>=4.8.0            # OpenCV图像处理库
numpy>=1.24.0                   # 数值计算库
pillow>=10.0.0                  # Python图像处理库

# 通信协议
paho-mqtt>=2.0.0                # MQTT客户端
paramiko>=3.0.0                 # SFTP/SSH客户端

# 配置与日志
pyyaml>=6.0                     # YAML配置文件解析

# 数据处理
scipy>=1.10.0                   # 科学计算（标定算法）

# ========================================================
# 可选依赖 - 根据平台选择
# ========================================================

# 嵌入式平台推理（ARM/RK3588）
# rknn-toolkit2                 # RKNN模型推理（仅ARM平台）

# ========================================================
# 开发依赖 - 仅开发环境需要
# ========================================================

# 代码质量
# pytest>=7.0.0                 # 单元测试框架
# pytest-cov>=4.0.0             # 测试覆盖率
# mypy>=1.0.0                   # 静态类型检查
# black>=23.0.0                 # 代码格式化
# flake8>=6.0.0                 # 代码风格检查
# pylint>=2.17.0                # 代码质量检查

# ========================================================
# 系统要求
# ========================================================
# Python >= 3.8
# 操作系统: Windows 10+, Ubuntu 18.04+, CentOS 7+
# 
# PC端（PyTorch）建议配置：
#   - CPU: Intel Core i5+ / AMD Ryzen 5+
#   - GPU: NVIDIA GTX 1650+ (支持CUDA)
#   - 内存: 8GB+
# 
# 嵌入式（RKNN）建议配置：
#   - SoC: RK3588 / RK3566
#   - 内存: 4GB+
# ========================================================

# ========================================================
# 安装说明
# ========================================================
# 
# 1. 创建虚拟环境（推荐）：
#    python -m venv venv
#    source venv/bin/activate    # Linux/Mac
#    venv\Scripts\activate       # Windows
# 
# 2. 安装依赖：
#    pip install -r requirements.txt
# 
# 3. GPU加速（可选，仅PC端）：
#    # 如需CUDA支持，请先安装对应版本的PyTorch
#    # 访问 https://pytorch.org 查看适合您系统的安装命令
# 
# 4. 验证安装：
#    python -c "import cv2, numpy, ultralytics, paho.mqtt, paramiko, yaml; print('所有依赖安装成功')"
# 
# ========================================================

